Re-work method to replace asm contents
The old method of replacing asm contents was buggy and slow on Emacs
26, causing many hangs when doing large replaces. This method is more
crude, but is much faster and less prone to subtle issues. With small
edits, the point may end up in the wrong place (as it is relying on
position rather than contents), but I think that's an acceptable loss.

Please file an issue if this causes issues for you. On Emacs 25, this
should strictly be better than the previous solution