Add support for Go
Go objdump produces its own output format which is dissimilar to GNU
objdump. The disassembly processor copies the Python approach, to
reformat the Go objdump output to be similar to GNU objdump. There are
some assumptions here about the output format of both packages, but it
seems to work well enough.